Add files via upload
This is the first view of the frontend design i am working on the backend and its other integration. Any suggestions and improvements will be of great value. 
I am making this as a Project "Theory of delay" right now just for one purpose in its execution mode as version 1.0 but in long run its a mathematical model for all the fields with custom ai that will help figure out the delay areas and calculate the indie task that can be done in that time quantum. And predictions of delay will help make better schedule areas and distribution in order to attaint the better efficiency. And maybe as a mathematical and ai computing future research scope too.

I will share all the ongoing work. I am adding the problem statements and all generic details. If anyone have got any ideas and wish to help or give suggestion. Your brother is open for all of that..  : )